Use nodeIP as the ip of grpc
