+++ pkg/utils/util.go
@@ -38,6 +38,7 @@ import (
 	"k8s.io/client-go/tools/record"
 	k8svol "k8s.io/kubernetes/pkg/volume"
 	"k8s.io/kubernetes/pkg/volume/util/fs"
+	"net"
 	"net/http"
 	"os"
 	"os/exec"
@@ -586,4 +587,33 @@ func WriteAndSyncFile(filename string, data []byte, perm os.FileMode) error {
 // Fsync is a wrapper around file.Sync(). Special handling is needed on darwin platform.
 func Fsync(f *os.File) error {
 	return f.Sync()
+}
+
+//GetNodeAddr get node address
+func GetNodeAddr(client kubernetes.Interface, node string, port string) (string, error) {
+	ip, err := GetNodeIP(client, node)
+	if err != nil {
+		return "", err
+	}
+	return ip.String() + ":" + port, nil
+}
+
+// GetNodeIP get node address
+func GetNodeIP(client kubernetes.Interface, nodeID string) (net.IP, error) {
+	node, err := client.CoreV1().Nodes().Get(context.Background(), nodeID, metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+	addresses := node.Status.Addresses
+	addressMap := make(map[v1.NodeAddressType][]v1.NodeAddress)
+	for i := range addresses {
+		addressMap[addresses[i].Type] = append(addressMap[addresses[i].Type], addresses[i])
+	}
+	if addresses, ok := addressMap[v1.NodeInternalIP]; ok {
+		return net.ParseIP(addresses[0].Address), nil
+	}
+	if addresses, ok := addressMap[v1.NodeExternalIP]; ok {
+		return net.ParseIP(addresses[0].Address), nil
+	}
+	return nil, fmt.Errorf("Node IP unknown; known addresses: %v", addresses)
 }
